+++ tests/testthat/test_simulating.R
@@ -52,13 +52,13 @@ test_that("Predictions are added", {
                                         explained_instance = X[3, ], 
                                         explained_var = "V1",
                                         size = 50)
-  local_dataset1 <- add_predictions(X, local_dataset, "regr.lm")
-  expect_output(add_predictions(X, local_dataset, "regr.lm"), regexp = NA)
+  local_dataset1 <- add_predictions(local_dataset, "regr.lm", X)
+  expect_output(add_predictions(local_dataset, "regr.lm", X), regexp = NA)
   expect_equal(length(local_dataset1), 6)
   expect_is(local_dataset1$data[[local_dataset1$target]], "numeric")
   expect_is(local_dataset1$target, "character")
-  local_dataset2 <- add_predictions(X, local_dataset, lm(V1 ~., data = X))
-  expect_output(add_predictions(X, local_dataset, lm(V1 ~., data = X)), regexp = NA)
+  local_dataset2 <- add_predictions(local_dataset, lm(V1 ~., data = X))
+  expect_output(add_predictions(local_dataset, lm(V1 ~., data = X)), regexp = NA)
   expect_equal(length(local_dataset2), 6)
   expect_is(local_dataset2$data[[local_dataset2$target]], "numeric")
   expect_is(local_dataset2$target, "character")